- ClientRegistration - use Expires header if no Contacts are present - Finish implementing RESIP_FIXED_POINT compiler flag (works for dum, stack, rutil and ares)
git-svn-id: https://svn.resiprocate.org/rep/resiprocate/main/resip@6414 ddefafc4-47db-0310-ae44-fa13212b10f2

+++ dum/ServerRegistration.cxx
@@ -62,7 +62,7 @@ ServerRegistration::accept(SipMessage& ok)
     ok.header(h_Contacts).push_back(contact);
     if(i->useQ)
     {
-      ok.header(h_Contacts).back().param(p_q)=i->q;
+      ok.header(h_Contacts).back().param(p_q)=(int)i->q;
     }
   }
